Move hooks into separate folder. Add show scrollbar hook
