+++ src/features/animals/Animals.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Section } from "lib/utils/useScrollIntoView";
+import { Section } from "lib/utils/hooks/useScrollIntoView";
 
 import { Chickens } from "./components/Chickens";
 import { GRID_WIDTH_PX } from "features/game/lib/constants";
