+++ src/features/hud/components/Menu.tsx
@@ -4,9 +4,8 @@ import { useActor } from "@xstate/react";
 import { Button } from "components/ui/Button";
 import { OuterPanel, Panel } from "components/ui/Panel";
 
-import { Section, useScrollIntoView } from "lib/utils/useScrollIntoView";
+import { Section, useScrollIntoView } from "lib/utils/hooks/useScrollIntoView";
 import * as Auth from "features/auth/lib/Provider";
-import { sync } from "features/game/actions/sync";
 import { Context } from "features/game/GameProvider";
 
 import { Share } from "./Share";
@@ -26,7 +25,7 @@ export const Menu = () => {
   const { authService } = useContext(Auth.Context);
   const { gameService } = useContext(Context);
   const [authState] = useActor(authService);
-  const [gameState, send] = useActor(gameService);
+  const [gameState] = useActor(gameService);
 
   const [menuOpen, setMenuOpen] = React.useState(false);
   const [scrollIntoView] = useScrollIntoView();
@@ -40,6 +39,7 @@ export const Menu = () => {
         window.location.href.includes("?")
           ? window.location.href.split("?")[0]
           : window.location.href
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       }?farmId=${authState.context.farmId!.toString()}`
     : "https://sunflower-land.com/play/";
