+++ src/lib/utils/hooks/useScrollIntoView.ts
@@ -1,5 +1,3 @@
-import { MutableRefObject, useEffect, useState } from "react";
-
 export enum Section {
   Crops = "crops",
   Water = "water",
